+++ test/test_um.rb
@@ -917,32 +917,28 @@ class QueueTest < UMBaseTest
     q = UM::Queue.new
     buf = []
 
-    f1 = Fiber.new do
+    machine.spin do
       buf << [1, machine.pop(q)]
-      machine.yield
     end
 
-    machine.schedule(f1, nil)
-
-    f2 = Fiber.new do
+    machine.spin do
       buf << [2, machine.pop(q)]
-      machine.yield
     end
-
-    machine.schedule(f2, nil)
 
     machine.snooze
     assert_equal [], buf
+    assert_equal 2, machine.pending_count
 
     machine.push(q, :foo)
     assert_equal 1, q.count
-    machine.sleep(0.02)
+    machine.snooze
+    assert_equal 1, machine.pending_count
     assert_equal [[1, :foo]], buf
 
     machine.push(q, :bar)
     assert_equal 1, q.count
 
-    machine.sleep(0.02)
+    machine.snooze
     assert_equal [[1, :foo], [2, :bar]], buf
     assert_equal 0, q.count
   end
@@ -969,7 +965,7 @@ class QueueTest < UMBaseTest
     end
     machine.schedule(f2, nil)
 
-    3.times { machine.snooze }
+    machine.snooze
 
     assert_equal [[1, :bar], [2, :foo]], buf.sort
     assert_equal 0, q.count
@@ -996,12 +992,12 @@ class QueueTest < UMBaseTest
     end
     machine.schedule(f2, nil)
 
-    machine.sleep 0.01
+    machine.snooze
 
     assert_equal [[1, :foo]], buf
     machine.push(q, :bar)
 
-    machine.sleep 0.01
+    machine.snooze
     assert_equal [[1, :foo], [2, :bar]], buf
   end