+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Improve `#snooze` to prevent op completion starvation
+
 # 2025-04-23 Version 0.6
 
 - Add `#periodically` for multishot timeout